producto ahora es eliminado del HD cuando hay forceDeletes

+++ app/Providers/ProductServiceProvider.php
@@ -15,15 +15,28 @@ class ProductServiceProvider extends ServiceProvider {
   public function boot()
   {
     Product::deleting(function($model){
-      $this->images = $model->images;
-      $this->id = $model->id;
+      if ($this->forceDeleting = $model->isForceDeleting())
+      {
+        $this->images = $model->images;
+        $this->id = $model->id;
+      }
     });
 
     Product::deleted(function($model){
-      if ($this->images) $this->images->each(function($image){
-        $image->delete();
-      });
-      return Storage::disk('public')->deleteDirectory("products/{$this->id}");
+      if ($this->forceDeleting)
+      {
+        if ($this->images) $this->images->each(function($image){
+          $image->delete();
+        });
+
+        if ($model->visits) $model->visits->each(function($visit){
+          $visit->delete();
+        });
+
+        if ($model->direction) $model->direction->delete();
+
+        return Storage::disk('public')->deleteDirectory("products/{$this->id}");
+      }
     });
   }
